+++ Provenance/zipkit/NSDate+ZKAdditions.h
@@ -9,7 +9,7 @@
 
 @interface NSDate (ZKAdditions)
 
-+ (NSDate *) zk_dateWithDosDate:(NSUInteger) dosDate;
-- (NSUInteger) zk_dosDate;
++ (NSDate *) zk_dateWithDosDate:(NSUInteger)dosDate;
+- (UInt32)      zk_dosDate;
 
 @end
